Add Wavelet Loss. Re-organize code. Add CLI tool.
