+++ .github/workflows/playwright.yaml
@@ -2,6 +2,7 @@ name: Playwright Tests
 on:
   push:
     branches: [ main, master ]
+permissions: read-all
 jobs:
   test:
     timeout-minutes: 60
