+++ .github/workflows/test.yaml
@@ -1,4 +1,5 @@
 on: push
+permissions: read-all
 jobs:
   test:
     strategy:
