+++ .github/workflows/wasm-bindgen-test.yaml
@@ -2,7 +2,7 @@ on:
   push:
     branches:
       - main
-
+permissions: read-all
 jobs:
   test:
     runs-on: ubuntu-latest
