+++ crates/lora-inspector-wasm/assets/js/lib.js
@@ -1,0 +1,4 @@
+import { parseSDKey } from "./moduleBlocks";
+
+
+export { parseSDKey };
