+++ crates/lora-inspector-wasm/assets/js/worker.js
@@ -528,3 +528,4 @@ async function getNetworkType(e) {
 
 	return loraWorker.network_type();
 }
+
