+++ crates/lora-inspector-wasm/index.html
@@ -41,6 +41,11 @@
     <!-- End Cloudflare Web Analytics -->
     <script src="/assets/js-lib/chartist.min.js"></script>
     <script src="/assets/js-lib/chartist-plugin-pointlabels.min.js"></script>
+		<style>
+		body {
+			background-color: #020f0f;
+		}
+		</style>
   </head>
   <body>
     <header class="home">
