+++ crates/lora-inspector/Cargo.toml
@@ -8,4 +8,11 @@ edition.workspace = true
 
 [dependencies]
 inspector = { path = "../inspector" }
-candle-core = { workspace = true }
+candle-core = { workspace = true, features = [] }
+clap = { version = "4.5.37", features = ["derive"] }
+serde.workspace = true
+serde_json = "1.0.140"
+indicatif = "0.17.11"
+
+[features]
+cuda = ["candle-core/cuda"]
